backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
langchain==0.1.0
langchain-ollama==0.0.1
langchain-google-genai==0.0.8
python-dotenv==1.0.1